Moved sample rate configuration to EEPROM.
Previously, we configured the sample rate entirely in sw. This
was convenient for the prototyping phase of the BattOr but complicates
the protocol for interfacing with software such as Chrome. Now there is
only one sample rate for UART and one for SD storage, and all of the
configuration parameters (i.e., timer, filter) for those sample rates.