Must sleep at least 0.6ms after digipot reset.
This resulted in a very strange bug where the digipot self test
was failing occasionally. The digipot takes at most 0.6ms to
reset before it can take any other commands. Therefore, I needed to
add this sleep after the reset to ensure the digipots are ready to
accept the next command.