Actually fixed bug with USART by writing all but one byte with DMA.
The issue turned out to be that the USART SPI DMA tx was not completing.
The reason why it was not completing was because the XMEGA's USART is
impossible to determine if the final byte transfer is complete when it was
transferred with DMA. To fix this problem we now transfer the last byte by
manually loading registers.