Two big changes to the firmware implementation.
(1) UART write switched to a much cleaner state machine implementation.
Which enabled (2) SD card stores UART frames.

This change makes it possible to have a large calibration frame on the UART.
It also switches to the SD card storing UART frames rather than raw samples
so they are faster to dump and if there is a change in the firwmare the old
samples can be interpreted correctly. It also improves accuracy because the
streaming and storage sampling happens with almost all of the same code
(except for the UART and filesystem).