Switched back to storing raw samples on the SD card.
Storing UART frames on the SD card is a problem because you end up
storing almost twice as much data as necessary because the UART frames
can be twice as long as the raw samples. Unfortunately, the SD card writes
at 10 ksps can not keep up with this rate. I also made an attempt to
make the SD write conditions similar to the streaming conditions so they
can share the same calibration data.