*** Significant change to how BattOr measures current ***
We had originally planned on using a mux to switch the input of the
ADC to GND for auto-calibration. However, this has proven to be problematic
because the mux has some leakage between the inputs. When it's connected to
GND, the output depends somewhat on the unconnected input from the sense
resistor.

To resolve this problem we have switched to leaving the mux connected to
the resistor at all times, and we only do auto-calibration on the ADC
itself (to account for temperature changes in the ADC). Given that we
added a bias voltage to the amplifier in v3.2.1, we should be able to
just rely on a factory calibration to account for that bias and the incorrect
gain value of the amplifier.