Added UART clock sync control message.
In order to get this to work it involved a few changes to the DMA
and UART code. Since sample write always writes a fixed size, we
removed the size argument. Also the sample size needed to be
updated because 50 samples was too short for handling the
control message. Also, I included a fix to the UART buffer
sizing which isn't necessary but it came to mind while I was
working on this commit.