Flush FTDI buffer immediatly after sync.
Chromium clock sync needs the FTDI's buffer to be flushed
as soon as the clock sync message hits it because Chromium
determines when the clock sync actually happeend based on
the round trip time of the sync message.

This required removing the unneccesary interrupt_disable()
in uart_tx_bytes() which would cause DMA interrupts to be
missed.