Fixed nasty SPI DMA bug by not using DMA for last RX byte in DMA transaction.
It turns out that it was a bad idea to try to use DMA to receive the last
byte of a SPI DMA transaction. For whatever reason this does not work all the
time, especially when there are other transactions going on the same channel
with the USART (even if they are not likely to be happening at the same time).
The problem was the last byte of the RX transaction would not complete.