Maxed out the UART TX download speed.
This makes it so downloading from the SD card uses the full bandwidth
of the UART channel (2 Mbit/s). The way I accomplished this was with three
main changes:

(1) Fix a bug with SPI DMA reads where the reads are missing one byte every
once in a while because the SPI clock was running faster than the DMA can
keep up with.

(2) Setup a double buffer for the preparation of a sample frame from
the SD card, and the transmission of a sample frame.

(3) Switched to using microsecond-scale delay for timeout when there is no UART
receiver.