Fixed #36 - Firmware UART TX doesn't pause when flow control occurs before TX starts
Added a state variable to track the flow control state even when the DMA UART TX is
not using the DMA channel. Also, the UART TX DMA init now sets the trigger
based on this state rather than forcing a trigger that will immediately start
the transaction even if the UART TX flow control has initiated a pause.