Simplified USART BAUD rate setting.
There were many instances of bit munging to set the baud rate all over
the firmware. Now we have one function to do this. Doing this also
helped me catch another smaller issue which was that the SPI baud
rate was being miscalculated because we were not using the USART
BAUD rate formula for SPI Master mode.