Fixed button self test.
The basic button driver broke the SRAM because I was missing a ~
in setting the direction of the button's GPIO pin so it made the SRAM's
CS into an input. Also I forgot to add the button self test to the main
self test routine. It is clearly getting too late for me.