Fixed #49 - Regression where INIT during reset causes UART to fail.
This regression came in 42fbb5ac, where I introduced a function for
setting UART baud rate. It turns out that in the UART init, setting
the baud rate in a function causes the UART to fail after reset. Instead
the baud rate setting function is now included as a static inline function.