Add real time clock timestamping for portable mode. (#52)
The way this works is the PC updates the BattOr's clock every time the battor sw is run. Since the BattOr is always powered on when connected to a test device, the internal ms timer just keeps the RTC time up to date.  When a portable trace is run the start of the trace is stored with the trace file in the fs.