Fixed bug where get count latency was at least 16 ms.
This was a subtle but nasty issue. First the FTDI's CTS is defaulted
to 0 (allow transmission), so toggling it should have gone to 1
then back to 0. Second, there was a syntax issue where PORTD.DIR should
have been PORTD.OUT in uart_set_rts(), this prevented the CTS line
from being toggled on the FTDI chip.