Abort SD card read over UART when bytes arrive over UART.
When bytes arrive over UART, that means the PC is trying
to run a new command, and thus abort the current SD card read.
This required changing the way that we signal the existance of bytes on
the UART. Instead of using the interrupt upper/lower half conecept, we
now just check to see if there are UART bytes pending to be read.

This change was needed because the battor_agent retries StopTracing when
it receives a corrupt message, and if the download continues the unread
data will crash the FTDI driver on Mac (which is also the platform that
drops lots of bytes).