Fixed #65: sample count command causes firmware reset.
The firmware was blocking until the UART transmission completed for the
reply to the sample count commant. In some cases the DMA transaction
would be held up by the flow control of the PC, and that would cause one
of the ADC DMA channels to overflow, leading to the reset.

This fix makes the sample count transmission asynchronous. Now an
interrupt at the end of the UART DMA transaction will trigger the flush
of the FTDI buffer, which was what we were blocking to wait for in the
first place.