[fw] Handle errors in uart rx frames.
When there was an error in a uart frame that was received from the host,
the firmware would just get stuck trying to read past the invalid frame.
This was becuase the parser was not designed to skip over uart rx bytes
that yeild an invalid frame.

This change significantly beefs up the uart rx parser. It is now robust
to any errors that I can imagine, as it properly skips over erronious bytes.
I also added testing code that should have been there in the first place to
ensure that the uart rx parser is working properly.